pyproject.toml
[project]
name = "a-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "braintrust>=0.5.4",
    "braintrust-langchain>=0.2.1",
    "google-genai>=1.61.0",
    "grandalf>=0.8",
    "langchain-core>=1.2.8",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "questionary>=2.0.0",
    "rich>=13.0.0",
    "tavily-python>=0.7.21",
]
